refactor: first part

+++ subgraphs/diva-staking/entities/Redemption.ts
@@ -1,4 +1,4 @@
-import { BigInt, Address, ethereum } from '@graphprotocol/graph-ts';
+import { BigInt, Address, ethereum, BigDecimal } from '@graphprotocol/graph-ts';
 import { uniqueEventId } from '@enzymefinance/subgraph-utils';
 import { Redemption } from '../generated/schema';
 import { Claim, Tranche } from '../utils/tranches';
@@ -16,7 +16,7 @@ export function createRedemption(
   let redemption = new Redemption(redemptionId(redeemer, event));
 
   redemption.redeemer = redeemer;
-  redemption.amounts = tranches.map<BigInt>((tranche) => tranche.amount);
+  redemption.amounts = tranches.map<BigDecimal>((tranche) => tranche.amount);
   redemption.trancheIds = tranches.map<i32>((tranche) => tranche.id as i32);
   redemption.firstClaimAmount = accruedRewards.firstClaimAmount;
   redemption.secondClaimAmount = accruedRewards.secondClaimAmount;
